Apply the same template to BungeeCord plugin.
