+++ Bungee/src/main/java/com/mizuledevelopment/bungee/jedis/JedisManager.java
@@ -1,0 +1,25 @@
+package com.mizuledevelopment.bungee.jedis;
+
+import lombok.Getter;
+import redis.clients.jedis.JedisPool;
+
+@Getter
+public class JedisManager {
+
+    private final JedisPool jedisPool;
+    private final String jedisChannel;
+
+    private final JedisSubscriber jedisSubscriber = new JedisSubscriber();
+    private final JedisPublisher jedisPublisher = new JedisPublisher();
+
+    public JedisManager(String host, int port, String jedisChannel, String jedisPassword) {
+        this.jedisPool = new JedisPool(host, port);
+        this.jedisChannel = jedisChannel;
+
+        if (jedisPassword != null)
+            this.jedisPool.getResource().auth(jedisPassword);
+
+        new Thread(() -> this.jedisPool.getResource().subscribe(this.jedisSubscriber, this.jedisChannel)).start();
+    }
+
+}
