+++ Bungee/src/main/java/com/mizuledevelopment/bungee/jedis/JedisPublisher.java
@@ -1,0 +1,11 @@
+package com.mizuledevelopment.bungee.jedis;
+
+import com.mizuledevelopment.bungee.BungeePlugin;
+
+public class JedisPublisher {
+
+    public void publishData(String message) {
+        BungeePlugin.getInstance().getJedisManager().getJedisPool().getResource().publish(BungeePlugin.getInstance().getJedisManager().getJedisChannel(), message);
+    }
+
+}
