+++ Node/src/main/resources/plugin.yml
@@ -1,3 +1,4 @@
 main: com.mizuledevelopment.node.NodePlugin
-version: ${project.version}
-name: Node
+name: PPS-Node
+version: ${git.commit.id.abbrev}
+author: MizuleDevelopment
